requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
sqlalchemy==2.0.43
pydantic==2.11.7
requests==2.32.5
python-dotenv==1.1.1
tenacity==9.1.2
xmind==1.2.0
